+++ lib/spack/docs/conf.py
@@ -97,6 +97,7 @@ sphinx_apidoc(
         "_spack_root/lib/spack/spack/test",
         "_spack_root/lib/spack/spack/package.py",
         "_spack_root/lib/spack/spack/new_installer_windows.py",
+        "_spack_root/lib/spack/spack/util/lock_windows.py",
     ]
 )
 sphinx_apidoc(
